\docType{class}
\name{bfSimpleOutputDF-class}
\alias{bfSimpleOutputDF-class}
\title{\code{bfSimpleOutputDF} is an S4 data.frame containing betfair tabular output.}
\description{
\code{bfSimpleOutputDF} is an S4 data.frame containing
betfair tabular output.
}
\examples{
getClass("bfSimpleOutputDF")
}
\author{
Vitalie Spinu
}
\seealso{
\code{\link{betfairly-package}}
\code{\link{bfSimpleOutput-class}}
\code{\link{bfSimpleOutputList}}
}
\keyword{class}

