\name{bfSimpleOutputDF-class}
\alias{bfSimpleOutputDF-class}
\alias{bfSimpleOutputDF}
\title{bfSimpleOutputDF is an S4 data...}
\description{\code{bfSimpleOutputDF} is an S4 data.frame containing betfair tabular output.}
\seealso{\code{\link{betfairly-package}} \code{\link{bfSimpleOutput-class}} \code{\link{bfSimpleOutputList}}}
\keyword{class}
\author{Vitalie Spinu}
\examples{getClass("bfSimpleOutputDF")}
\section{Extends}{\code{\linkS4class{data.frame}}, \code{\linkS4class{bfSimpleOutput}}}

